feat: Move image dependency to app folder
