+++ packages/core/app/components/search/SearchProductItem/SearchProductItem.tsx
@@ -5,7 +5,7 @@ import {
   useSearch,
 } from '@faststore/ui'
 
-import { Image } from 'src/components/ui/Image'
+import { Image } from 'app/components/ui/Image'
 import { useFormattedPrice } from 'src/sdk/product/useFormattedPrice'
 import { useProductLink } from 'src/sdk/product/useProductLink'
 import type { ProductSummary_ProductFragment } from '@generated/graphql'
