+++ packages/core/app/components/ui/Gift/Gift.tsx
@@ -7,7 +7,7 @@ import {
 } from '@faststore/ui'
 
 import type { CartItem as ICartItem } from 'app/sdk/cart'
-import { Image } from 'src/components/ui/Image'
+import { Image } from 'app/components/ui/Image'
 import { useFormattedPrice } from 'src/sdk/product/useFormattedPrice'
 
 export type Props = GiftProps & {
