+++ packages/core/app/components/ui/Image/loader.ts
