+++ packages/core/src/components/product/ProductCard/ProductCard.tsx
@@ -9,7 +9,7 @@ import { gql } from '@generated'
 import type { ProductSummary_ProductFragment } from '@generated/graphql'
 import { ImageProps } from 'next/image'
 import NextLink from 'next/link'
-import { Image } from 'src/components/ui/Image'
+import { Image } from 'app/components/ui/Image'
 import { useFormattedPrice } from 'src/sdk/product/useFormattedPrice'
 import { useProductLink } from 'src/sdk/product/useProductLink'
